+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Stencil Changelog
+
+## Master
+
+### Bug Fixes
+
+- Variables (`{{ variable.5 }}`) that reference an array index at an unknown
+  index will now resolve to `nil` instead of causing a crash.
+  [#72](https://github.com/kylef/Stencil/issues/72)
+
+
 ## 0.6.0
 
 ### Enhancements